Refactor `count` -> `estimated_document_count`
PyMongo `~=4.0` deprecates `Cursor.count` and `Collection.count` in favor of more explicit
`Collection.count_documents` and `Collection.estimated_document_count`. Both `Cursor.count`
and `Collection.count` in PyMongo `~=3.0` used the `count` MongoDB command, which would
potentially provide inaccurate results when used without a query predicate, in a transaction,
or on a sharded cluster. When running without a predicate, the count was returned based
on the collection's metadata - fast, but not necessarily correct.

The more explicit `count_documents` and `estimated_document_count` allow the client to
specify the expected operation. `Collection.count_documents` uses an aggregate query with
`$group` + `$sum` to provide counts in `~=4.0,<=4.10.1`.

`Collection.estimated_document_count` in `>=4.0,<4.2` would select between one of two
methods. When running against MongoDB 4.9+, `$collStats` was used to provide the collection
count, and the `count` command was used when running against MongoDB <4.9. Given some issues
with the 5.0.0-5.0.8 Stable API spec, PyMongo `>=4.2,<=4.10.1` reverted the `$collStats`
improvements and opted to use the `count` command universally for `estimated_document_count`,
technically making it compatible with the original `count` command.

We might want to evaluate adding a more accurate `count` aggregation method in the future,
but for now we can rely on the existing and well tested `count` command for our integration.